+++ src/components/@layout/Layout.tsx
@@ -2,7 +2,7 @@ import { Button, Stack, styled } from "@mui/material";
 import { Outlet, useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 import SideNavbar from "@/components/@common/SideNavbar";
-import useAuthStorage from "@/hooks/useAuthStorage";
+import useAuthStorage from "@/hooks/auth/useAuthStorage";
 import RoutePath from "@/routes/routePath";
 
 export default function Layout() {
